Switch logs to uint64 for very large simulations.

+++ stochastic_matching/simulator/age_based.py
@@ -321,10 +321,10 @@ class FCFM(QueueStateSimulator):
     >>> sim = sm.FCFM(sm.Cycle(rates=[3, 4, 5]), number_events=1000, seed=42, max_queue=10)
     >>> sim.run()
     >>> sim.logs # doctest: +NORMALIZE_WHITESPACE
-    {'trafic': array([125, 162, 213], dtype=uint32),
+    {'trafic': array([125, 162, 213], dtype=uint64),
     'queue_log': array([[838, 104,  41,  13,   3,   1,   0,   0,   0,   0],
        [796, 119,  53,  22,   8,   2,   0,   0,   0,   0],
-       [640, 176,  92,  51,  24,   9,   5,   3,   0,   0]], dtype=uint32),
+       [640, 176,  92,  51,  24,   9,   5,   3,   0,   0]], dtype=uint64),
     'steps_done': 1000}
 
     Unstable diamond (simulation ends before completion due to drift).
@@ -332,11 +332,11 @@ class FCFM(QueueStateSimulator):
     >>> sim = FCFM(sm.CycleChain(rates=[1, 1, 1, 1]), number_events=1000, seed=42, max_queue=10)
     >>> sim.run()
     >>> sim.logs # doctest: +NORMALIZE_WHITESPACE
-    {'trafic': array([34, 42,  7, 41, 36], dtype=uint32),
+    {'trafic': array([34, 42,  7, 41, 36], dtype=uint64),
     'queue_log': array([[127,  70,  22,  26,  29,  12,  23,  15,  10,   5],
            [327,   8,   3,   1,   0,   0,   0,   0,   0,   0],
            [322,  12,   4,   1,   0,   0,   0,   0,   0,   0],
-           [106,  80,  65,  28,  31,  15,   4,   2,   6,   2]], dtype=uint32),
+           [106,  80,  65,  28,  31,  15,   4,   2,   6,   2]], dtype=uint64),
     'steps_done': 339}
 
     A stable candy (but candies are not good for greedy policies).
@@ -345,7 +345,7 @@ class FCFM(QueueStateSimulator):
     ...            number_events=1000, seed=42, max_queue=25)
     >>> sim.run()
     >>> sim.logs # doctest: +NORMALIZE_WHITESPACE
-    {'trafic': array([24, 17,  2, 23, 33, 12, 13], dtype=uint32),
+    {'trafic': array([24, 17,  2, 23, 33, 12, 13], dtype=uint64),
     'queue_log': array([[ 24,  32,  45,  38,  22,  43,  31,  34,  20,   3,   0,   0,   0,
               0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
            [291,   1,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
@@ -360,7 +360,7 @@ class FCFM(QueueStateSimulator):
               0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
            [232,  33,  16,   4,   6,   1,   0,   0,   0,   0,   0,   0,   0,
               0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0]],
-          dtype=uint32),
+          dtype=uint64),
     'steps_done': 292}
     """
